Implement suggested review improvements
